Add `FixedInteger` trait
Adds a trait specifically for `Uint`-like fixed-width integers whose
size is known at compile time.

Also adds a blanket impl of the `Integer` trait for all `FixedInteger`
types, and adds a default impl of `is_odd` to `Integer`.
